Revert "Refactor: extract timeline builder, fix review feedback"
This reverts commit 2e6b2db85f37f1cec877170678020d046fc272cb.
