+++ remotion/src/DailyCard.tsx
@@ -19,7 +19,6 @@ import {
   Item,
   DailyCardProps,
   DATE_FRAMES,
-  OPENING_FRAMES,
   computeTotalFrames,
   computeScaleFactor,
   wordFrames,
@@ -30,8 +29,7 @@ import {
   DEFAULT_KEY_FACTS_FADER_CONFIG,
   DEFAULT_COUNCIL_FADER_CONFIG,
 } from "./timing";
-import { GraphCanvas, resolveGraphCamera, getTopicColorForFrame } from "./graph/GraphCanvas";
-import { buildGraphTimeline } from "./graph/timeline";
+import { GraphCanvas, buildGraphTimeline, resolveGraphCamera, OPENING_FRAMES, getTopicColorForFrame } from "./graph/GraphCanvas";
 import { GLBFader } from "./GLBFader";
 import { buildFaderSceneBounds } from "./fader";
 import { CANVAS_SIZE } from "./graph/layout";
